forgot to change this

+++ src/main/java/com/rabbitMq/rabbitmqscheduler/DTO/UserTransferOptions.java
@@ -17,6 +17,6 @@ public class UserTransferOptions {
     private Integer concurrencyThreadCount;
     private Integer parallelThreadCount;
     private Integer pipeSize;
-    private Integer chunkSize;
+    private Long chunkSize;
 
 }
